Tackle the case of bare GIT repositories for PS1 terminal
When entering in a bare repository, git_status_reference failed because git
status cannot be done on a bare repository with a fatal error. A simple check
on the value of core.bare is done, default being false for a normal GIT work
repository. If the current repository is a GIT bare one, a simple message
(bare-repo is printed for PS1).