Move all the compilation utilities into the unique pg_compile
pg_compile integrates now options to copy source code to another folder and
execute vpath and normal installations. The following options are added:
- i to define an input folder containing source code
- o to define an output folder used for compilation
- n flag to to a normal installation
- v flag to do a vpath installation
In the case of a normal installation, the code is copied from input to output
then compiled. For a VPATH, a new output folder is created and the configure
located in input folder is used for configuration.
In consequence, utilities makepostgres and makevpath are removed.