Perform GIT completion script through PROMPT_COMMAND instead of PS1
Doing that makes performance better as PS1 is only designed to print things and
not to launch internal commands. So what is done now is to recompile the main
function of script ~/bin/git_completion each time the prompt is update and
generate an environment variable called GIT_PS1_COMPLETION. The this variable
is used by PS1 to update screen directly.

Update some ~/ to $HOME at the same time in .bashrc