Addition of pg_binarchive, script to automatically install PG stable versions
All those installation are included in HOME_STABLE_BIN=$HOME/bin/stable and
are not included in $PATH by default so they should be called with full path
to work properly. It is also important to remember that LD_LIBRARY_PATH should
be updated to a value indicating compatible libs if necessary.