Add support for ccache and improve pg_configure for non-copy mode
When pg_compile is used in non-copy mode, the root tree of the git folder
where call is being done from is used for the configure phase. This allows
calls of pg_compile to be done inside a postgres git tree.

Also, support for ccache is added with this commit with the addition of
a new option -e that needs to specify a sub ccache name to designate a
folder in HOME_CCACHE_DIR. This option is not mandatory and let at the
convenience of the user.