Support of custom parameters for postgresql.conf and recovery.conf
With initializing a new Postgres node, user can now specify a custom set
of parameters for postgresql.conf or recovery.conf which are uploaded
directly when creating the node. The files containing custom parameters
are located in ~/etc/postgres.d, aka $HOME_POSTGRES_ETC.

Also the following additions are done for postgres_node_init:
- Addition of option -c to specify a suffix for postgresql.conf. The file
fetched from etc folder is postgresql.$SUFFIX.conf and is added at the
bottom of postgresql.conf automatically, allowing user to override some
parameter values. The parameters added in those files should not interact
with the ones automatically generated in postgres_node_init (mainly port
and archive_command).
- -r option does the same for recovery.conf. Generic parameters are in
this case restore_command and primary_conninfo so they should not, but
could be overriden to some more specific values. In this case the file
automatically fetched from etc is called recovery.$SUFFIX.conf.

Users are able to add new custom configuration files by adding them
in $HOME_POSTGRES_ETC with the correct format name and with custom
parameters, what can be useful for debugging purposes. For the time
being all the files that have not the suffix dev are ignored, making
$HOME_POSTGRES_ETC half private/public. A different folder could have
been used to differentiate both cases but I didn't think it was worth
the pain and would make script code unnecessarily complicated.

pg_start is not extended yet with those new options, it will be with
a subsequent commit.