Remove concept of separate path for build and dev in pg_compile
This removes a couple of things like support for vpath build but it
makes it more extensible for future things related to Postgres, like
the addition of an option for doc compilation. This code became a bit
a mess btw with the idea of normal, vpath and copy-only install without
real use.

This is the conclusion of months of development with the former script,
that had a set of options I am not really using on a daily basis.