Support automatic merge of remote branches for pg_compile
Prior doing that, pg_compile was simply checking out a given branch
and sticked to it, resulting in a code not in sync with a given remote
if the branch already exists in git repository. Hence, this commit adds
a new -o option for pg_compile to define a remote link

pg_binarchive is updated as well to use that, and fetch the latest
commit of a branch checked out automatically as what is targetted is
the latest point of a branch and not a given commit.