Replace femacs by git-open and use EDITOR as default editor
Compared to femacs that was scanning all the files in a folder, even
the ones that were not part of a git repository, git-open only scans
files in the code tree. After months spent with femacs, I also noticed
that I used it only inside git repositories, so git-open is really going
to make life easier.

using EDITOR has the advantage to let the user decide the editor they
want to give priority to in their environment without forcing them to use
emacs. For the same reason, core.editor is moved from ~/.gitconfig to
configuration template of spatch where users are free to use the editor
they want. git-open makes use of this facility, future scripts that would
need to interact with an editor should do the same.