Append new commands in PROMPT_COMMAND instead of enforcing its value
It happens that OSX uses PROMPT_COMMAND to enforce the opening directory
of a new tab to be the same one as the original tab from which new tab is
opened. Appending at the bottom of PROMPT_COMMAND the new instructions
seems to do the trick, and it worked when tried on OSX, Archlinux, Fedora
when tested.