Remove spaces from PROMPT_COMMAND to save space
git prompt allows a high level of customization with the presence of
a 3rd argument that can be used with __git_ps1 to specify the format
of the output string of printf. In the case of this environment,
'(%s%s)' is simply used. While at it, a mention of GIT_PS1_STATESEPARATOR
is added in spatch template to let the user know that it is possible to
personalize separator between branch name and statuses. This is
undocumented in the git prompt script.