Remove --host setting in pg_compile formerly used for MinGW-w64
It happens that commit fc75250 in Postgres upstream is breaking compilation
for old versions of MinGW-w64 (actually a snapshot dating of 2011 does not
work) because of a gai_strerror undefined. Trying to build the code with
newer versions of MinGW fails as well caused by invalid references to
_set_invalid_parameter_handler in stdlib.h when running ./configure (note
that this happens down to 4.8.1 findable on SourceForce). Hence this forces
pg_compile to fallback to vanilla MinGW where compilation works, under
two conditions though:
- #define stat _stat in sys/stat.h needs to be commented out
- #define DllMain in winbase.h needs to be commented out
At least with this patched version of MinGW things are working.