Add support for Makefile.custom in pg_compile
With the new option -s given, a user can specify a flag that is used
to select a custom Makefile named as Makefile.flag.custom in
$HOME_POSTGRES_ETC. The default is "dev", anc contains the CFLAGS
options used until now in pg_compile. This is not the way of doing
really recommended by Postgres, so user should really be aware of what
he does when using this facility.