Introduce FILES in spatch
FILES is a file containing a list of entries as follows:
filename:path/to/filename

spatch can use entries in FILES and create soft links as wanted by the
user. This is more flexible than the old system where the file list
of spatch was hardcoded in the utility itself. Now user can take some
configuration files in .examples and insert them in his own configuration
profile.