Correct inconsistent behaviors of spatch
The following things are changed:
- Error out if FILES is not found when getting a file list. This will
prevent the use of corrupted configuration sets. This area of the code
is still far from being perfect as there is no protection if either a
file or if subfolders of a link are used.
- Remove only existing configurations when calling command conf. The
older behavior was to reset unconditionnally as files to link were
hardcoded and not explicitely listed in FILES.
- Leave conf-reset if there is no current configuration.