Fix file profile removal for spatch on NFS systems
On NFS systems (aka Windows), ln works by creating a copy of the file
linked instead of really create a link to it, something that made spatch
not work properly when resetting a profile as check was done only thinking
that the profile file was linked. The logic has been changed for MINGW
environments to use a check on a file. It may be possible to change ln -s
to cp when setting a profile to decouple the behavior of ln, but it seems
unlikely that the maintainers of ln will create an implementation for NFS
either way, as it would require recreating a complete new thing. In my
opinion they are right to do so, so let's just bother with some additional
logic only when removing a profile for simplicity.