Make effective forced switch of SHELL all the time
When using pg_compile -c, Makefile.global is generated but unfortunately
the forced switch of variable SHELL to "bash" that is necessary to install
binaries in msysgit was not done, making this command followed by for
example a simple make check failing. This commit ensures that the switch
is done on MINGW* environments all the time, for any modes of pg_compile.