Allow titles of TERM tabs to be set dynamically
By using a method abusing of the DEBUG signal, the titles of terminal
tabs are now updated depending on the last command typed on it. A special
is made if EDITOR is used to replace the title with only the file name
used. This could be customized further, let's see how I can get used to
that first.