Update mutt parameters to always show mailboxes
Previous settings made the mailboxes show up only if there was new email,
but that's really unpractical as sometimes there is a need to have a look
at the exiting messages without any new ones.

Tweak as well shortcuts for sidebar when opening and switching mailboxes
so as they do not depend anymore on new mails.