Extend pg_compile to allow build with different source folders
This has been removed as of 565e5ac, but after more review of the code
it happens that there were still some leftovers to allow the script to
work with a different source folder than the original.  This is useful
to test.sh in pg_upgrade as this requires the old source folder and
binaries to be able to work across major upgrades.