Simplify pg_compile and pg_binarchive for worktree integration
pg_compile loses its options -b, -t and -o to control the git commit to
use as a base when compiling, as it has proved pretty useless to me for
the last couple of years.  This simplifies in consequence pg_binarchive,
so as it uses as source path the worktree repository.