Remove support for src/Makefile.custom in Postgres
PostgreSQL allows a developer to use src/Makefile.custom to pass down
custom make flags after ./configure.  While not recommended, this has
proved to be useful in some cases when bypassing ./configure failures.

This facility is useless with the meson builds, and CFLAGS/LDFLAGS can
already be handled by pg_compile via HOME_POSTGRES_{CFLAGS,LDFLAGS}, so
let's remove support for Makefile.custom in this set of scripts and rely
on the HOME_POSTGRES_* environment variable instead, so as both meson
and ./configure use the same set of flags.