Remove support for HOME_ENV_NAME in MinGW
These have not been used in ages for development on Windows, where the
CI and the buildfarm are good enough these days.  The remaining tweaks
for HOME_ENV_NAME concern Darwin and Linux, as detected by `uname`.

Some comments are cleaned up on the way, as are some non-active links.